# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: contact
    description: Contact form submissions
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit a contact enquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactRequest"
      responses:
        "201":
          description: Submission saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/submissions:
    get:
      operationId: getSubmissions
      tags: [admin]
      summary: Get all contact submissions (admin only)
      parameters:
        - name: password
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ContactRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        college:
          type: string
        projectType:
          type: string
        serviceType:
          type: string
          enum: [project_only, project_internship, internship_only]
        message:
          type: string
      required:
        - name
        - phone
        - email
        - college
        - projectType
        - serviceType

    ContactResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        id:
          type: integer
      required:
        - success
        - message
        - id

    Submission:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        college:
          type: string
        projectType:
          type: string
        serviceType:
          type: string
        message:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - phone
        - email
        - college
        - projectType
        - serviceType
        - createdAt

    SubmissionsResponse:
      type: object
      properties:
        submissions:
          type: array
          items:
            $ref: "#/components/schemas/Submission"
        total:
          type: integer
      required:
        - submissions
        - total

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error