# lib/api-zod/src/generated/types/contactRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContactRequestServiceType } from "./contactRequestServiceType";

export interface ContactRequest {
  name: string;
  phone: string;
  email: string;
  college: string;
  projectType: string;
  serviceType: ContactRequestServiceType;
  message?: string;
}