# lib/api-zod/src/generated/types/submission.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Submission {
  id: number;
  name: string;
  phone: string;
  email: string;
  college: string;
  projectType: string;
  serviceType: string;
  message?: string;
  createdAt: Date;
}