# lib/api-zod/src/generated/types/submissionsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Submission } from "./submission";

export interface SubmissionsResponse {
  submissions: Submission[];
  total: number;
}